Migrated unit tests for FilterAllowedOutgoingContentAttribute and MediaController
