+++ src/Umbraco.Tests.UnitTests/Umbraco.Web.BackOffice/Controllers/MediaControllerUnitTests.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
-using System.Web.Http;
 using Moq;
 using NUnit.Framework;
 using Umbraco.Core.Models;
 using Umbraco.Core.Models.Entities;
 using Umbraco.Core.Models.Membership;
 using Umbraco.Core.Services;
-using Umbraco.Tests.TestHelpers.Entities;
-using Umbraco.Web.Editors;
+using Umbraco.Tests.Common.TestHelpers.Entities;
+using Umbraco.Web.BackOffice.Controllers;
+using Umbraco.Web.Common.Exceptions;
 
 namespace Umbraco.Tests.Web.Controllers
 {
@@ -32,7 +32,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act
-            var result = MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, 1234);
+            var result = MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, 1234);
 
             //assert
             Assert.IsTrue(result);
@@ -55,7 +55,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act/assert
-            Assert.Throws<HttpResponseException>(() => MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, 1234));
+            Assert.Throws<HttpResponseException>(() => MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, 1234));
         }
 
         [Test]
@@ -78,7 +78,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act
-            var result = MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, 1234);
+            var result = MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, 1234);
 
             //assert
             Assert.IsFalse(result);
@@ -98,7 +98,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act
-            var result = MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, -1);
+            var result = MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, -1);
 
             //assert
             Assert.IsTrue(result);
@@ -120,7 +120,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act
-            var result = MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, -1);
+            var result = MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, -1);
 
             //assert
             Assert.IsFalse(result);
@@ -140,7 +140,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act
-            var result = MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, -21);
+            var result = MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, -21);
 
             //assert
             Assert.IsTrue(result);
@@ -162,7 +162,7 @@ namespace Umbraco.Tests.Web.Controllers
             var entityService = entityServiceMock.Object;
 
             //act
-            var result = MediaController.CheckPermissions(new Dictionary<string, object>(), user, mediaService, entityService, -21);
+            var result = MediaController.CheckPermissions(new Dictionary<object, object>(), user, mediaService, entityService, -21);
 
             //assert
             Assert.IsFalse(result);
